Feat: Refactor package structure and update imports to improve code organization
