+++ Backend/gestao-projetos/src/main/java/com/gestaoprojetos/service/StatusServiceIMP.java
@@ -4,6 +4,7 @@ package com.gestaoprojetos.service;
 import com.gestaoprojetos.exception.BadRequestException;
 import com.gestaoprojetos.exception.ResourceNotFoundException;
 import com.gestaoprojetos.model.Status;
+import com.gestaoprojetos.repository.BasicRepositoryIMP;
 import com.gestaoprojetos.repository.StatusRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
